requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-CORS==5.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-genai==1.5.0
grpcio==1.70.0
gspread==6.2.0
gunicorn==23.0.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
protobuf==6.30.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
